apps/backend/requirements.txt
fastapi==0.128.4
uvicorn==0.40.0
python-multipart==0.0.22
pydantic==2.12.5
pydantic-settings==2.12.0
tinydb==4.8.2
litellm==1.81.8
markitdown[docx]==0.1.4
pdfminer.six==20260107
python-docx==1.2.0
playwright==1.58.0
python-dotenv==1.2.1
supabase==2.15.1
PyJWT[crypto]==2.10.1
